+++ app/react-native/src/preview/components/OnDeviceUI/navigation/visibility-button.tsx
@@ -8,8 +8,8 @@ interface Props {
 const Touchable = styled.TouchableOpacity({
   backgroundColor: 'transparent',
   position: 'absolute',
-  right: '8',
-  bottom: '12',
+  right: 8,
+  bottom: 12,
   zIndex: 100,
 });