+++ examples/cra-ts-kitchen-sink/src/components/Button.stories.tsx
@@ -1,12 +1,25 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import Button from './Button';
+import { radios } from '@storybook/addon-knobs';
+import Button, { Type } from './Button';
 
-storiesOf('Button', module).add(
-  'simple button',
-  () => <Button onClick={action('button clicked')}>OK</Button>,
-  {
-    info: { inline: true },
-  }
+export default {
+  title: 'Docgen|Button',
+  component: Button,
+};
+
+type Story = () => any;
+
+export const simpleButton: Story = () => {
+  const x = 0;
+  return <Button onClick={action('button clicked')}>OK {x}</Button>;
+};
+
+const typeOptions = {
+  Default: 'default',
+  Action: 'action',
+};
+
+export const withType = () => (
+  <Button type={radios('Type', typeOptions, typeOptions.Default) as Type}>Label</Button>
 );